Adding Community Video Tutorials Section
Hello Piotr and team, you did an extraordinary work! I've created a video in which I'm using your products (you can check it here  - https://youtu.be/dZmjyMfJnCU?t=334) I thought it can help someone and I've add it in the readme.md. Feel free to approve it, edit it or remove it. Either way keep up the good work!